requirements.txt
allure-behave==2.6.0
allure-python-commons==2.6.0
altgraph==0.10.2
Appium-Python-Client==0.52
arrow==0.13.1
atomicwrites==1.3.0
attrs==17.4.0
backports-abc==0.5
backports.functools-lru-cache==1.5
bdist-mpkg==0.5.0
behave==1.2.6
behave2cucumber==1.0.3
boto==2.49.0
boto3==1.10.14
botocore==1.13.14
browserstack-local==1.2.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
config==0.4.0
docutils==0.15.2
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
futures==3.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
linecache2==1.0.0
logger==1.4
macholib==1.5.1
MarkupSafe==1.1.1
matplotlib==1.3.1
modulegraph==0.10.4
more-itertools==5.0.0
nose==1.3.7
numpy==1.16.2
parse==1.8.2
parse-type==0.4.2
pathlib2==2.3.3
pdfkit==0.6.1
pluggy==0.9.0
psutil==5.7.2
py==1.8.0
py2app==0.7.3
pyobjc-core==5.1.2
pyobjc-framework-Cocoa==5.1.2
pyobjc-framework-CoreData==5.1.2
pyobjc-framework-Quartz==5.1.2
pyparsing==2.0.1
pytest==4.3.1
pytest-dotenv==0.4.0
python-dateutil==2.8.0
python-decouple==3.3
python-dotenv==0.14.0
pytz==2013.7
requests==2.21.0
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.89
s3transfer==0.2.1
s3utils==0.6.1
sauceclient==1.0.0
scandir==1.10.0
schedule==0.6.0
selenium==3.141.0
singledispatch==3.4.0.3
six==1.11.0
slackclient==1.3.1
tornado==5.1.1
traceback2==1.4.0
typing==3.7.4.3
urllib3==1.24.1
virtualenv==15.1.0
websocket-client==0.54.0
Werkzeug==0.15.1
xattr==0.6.4
zope.interface==4.1.1
